+++ tests/functional/shared_tests_instances/single_layer_tests/transpose.cpp
@@ -30,7 +30,7 @@ using namespace LayerTestsDefinitions;
 namespace {
 
 const std::vector<InferenceEngine::Precision> netPrecisions = {
-    InferenceEngine::Precision::FP32,
+    InferenceEngine::Precision::FP16,
 };
 
 // MCM instantiation
